Radial Gas - first working version Solves the transient with radial symmetry as a finite difference problem. This will be much more general than the Ei solution.
I use the method of lines to make it a set of odeints. Then I use canned python odeint
Very important to set a small stepsize at beginning and to increase max iters from 500.

The hardest part is the origin. I exclude the origin from the grid and I set up a boundary condition that fixes the Pressure derivative as we approach the origin.

An irregular grid is coded but doesnt work, I am missing some kind of Jacobian.